Try to add stacked/tabbed titles to `_NET_FRAME_EXTENTS`
Add the area occupied by the title bars for stacked/tabbed containers to
the top frame size reported in `_NET_FRAME_EXTENTS`.

This change is not 100% correct, because these decorations are not
actually drawn into the X11 frame window to which the client window is
reparented, therefore software like `picom` won't be able to handle
those title bars anyway.

Signed-off-by: Sergey Vlasov <sigprof@gmail.com>